requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
daiquiri==3.0.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
ldap3==2.9.1
Levenshtein==0.26.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
packaging==24.2
pendulum==3.0.0
pillow==11.1.0
pip==25.0.1
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.26.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.11.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
thefuzz==0.22.1
time-machine==2.16.0
tomli==2.2.1
typer==0.15.1
typer-slim==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.45.1
zstandard==0.23.0
